Desafio 6 - UI - HUD
Creación del HUD
Creación UI / Portada / Menú

+++ Desafio_1_Hirschfeld/Assets/Scripts/Enemigo/Herir.cs
@@ -6,15 +6,15 @@ public class Herir : MonoBehaviour
 {
     // Variables a configurar desde el editor
     [Header("Configuracion")]
-    [SerializeField] float puntos = 5f;
+    [SerializeField] int puntos = 5;
 
     private void OnCollisionEnter2D(Collision2D collision)
     {
         if (collision.gameObject.CompareTag("Player"))
         {
             Jugador jugador = collision.gameObject.GetComponent<Jugador>();
-            jugador.ModificarVida(-puntos);
-            Debug.Log(" PUNTOS DE DAÑO REALIZADOS AL JUGADOR "+ puntos);
+                jugador.ModificarVida(-puntos);
+                Debug.Log(" PUNTOS DE DAÑO REALIZADOS AL JUGADOR "+ puntos);
         }
     }
 }
